backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   starlette
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
distro==1.9.0
    # via openai
fastapi==0.120.3
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.11.1
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.0.3
    # via -r requirements.in
langchain-classic==1.0.0
    # via -r requirements.in
langchain-core==1.0.2
    # via
    #   langchain
    #   langchain-classic
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-openai==1.0.1
    # via -r requirements.in
langchain-text-splitters==1.0.0
    # via langchain-classic
langgraph==1.0.2
    # via langchain
langgraph-checkpoint==3.0.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.2
    # via langgraph
langgraph-sdk==0.2.9
    # via langgraph
langsmith==0.4.38
    # via
    #   langchain-classic
    #   langchain-core
openai==2.6.1
    # via
    #   -r requirements.in
    #   langchain-openai
orjson==3.11.4
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.11.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
pydantic==2.12.3
    # via
    #   -r requirements.in
    #   fastapi
    #   langchain
    #   langchain-classic
    #   langchain-core
    #   langgraph
    #   langsmith
    #   openai
pydantic-core==2.41.4
    # via pydantic
python-dotenv==1.2.1
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   langchain-classic
    #   langchain-core
regex==2025.10.23
    # via tiktoken
requests==2.32.5
    # via
    #   langchain-classic
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.44
    # via langchain-classic
starlette==0.49.1
    # via fastapi
tenacity==9.1.2
    # via langchain-core
tiktoken==0.12.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via -r requirements.in
xxhash==3.6.0
    # via langgraph
zstandard==0.25.0
    # via langsmith
